+ Added main code for level creating, player creating and some systems implementations + Added to SimpleComponent with initializing of containing type + Added 'builder' pattern as SetMember in UserComponent when it claimed * Refactored custom level creating * Refactored Entity & Component base classes

+++ Game/src/Utility/Math/Vectors/Vectors.cpp
@@ -53,7 +53,7 @@ sf::Vector2f randUnitVector() {
 
 sf::Vector2f fabs(sf::Vector2f A)
 {
-    return sf::Vector2f { fabs(A.x), fabs(A.y) };
+    return sf::Vector2f { fabsf(A.x), fabsf(A.y) };
 }
 
 bool operator< (sf::Vector2f A, sf::Vector2f B)
